backend/ai/requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
python-dotenv
assemblyai
supabase
transformers
torch
